// src/API.ts
export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type Like = {
  __typename: "Like",
  id: string,
  userID: string,
  postID: string,
  User?: User | null,
  Post?: Post | null,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
  owner?: string | null,
};

export type User = {
  __typename: "User",
  id: string,
  name: string,
  email?: string | null,
  username?: string | null,
  bio?: string | null,
  website?: string | null,
  nofPosts: number,
  nofFollowers: number,
  nofFollowings: number,
  image?: string | null,
  Posts?: ModelPostConnection | null,
  Comments?: ModelCommentConnection | null,
  Likes?: ModelLikeConnection | null,
  Followers?: ModelUserFollowConnection | null,
  Followings?: ModelUserFollowConnection | null,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
  owner?: string | null,
};

export type ModelPostConnection = {
  __typename: "ModelPostConnection",
  items:  Array<Post | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type Post = {
  __typename: "Post",
  id: string,
  createdAt: string,
  type: string,
  description?: string | null,
  location?: string | null,
  image?: string | null,
  images?: Array< string > | null,
  video?: string | null,
  nofComments: number,
  nofLikes: number,
  userID: string,
  User?: User | null,
  Likes?: ModelLikeConnection | null,
  Comments?: ModelCommentConnection | null,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
  owner?: string | null,
};

export type ModelLikeConnection = {
  __typename: "ModelLikeConnection",
  items:  Array<Like | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type ModelCommentConnection = {
  __typename: "ModelCommentConnection",
  items:  Array<Comment | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type Comment = {
  __typename: "Comment",
  id: string,
  createdAt: string,
  comment: string,
  userID: string,
  postID: string,
  User?: User | null,
  Post?: Post | null,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
  owner?: string | null,
};

export type ModelUserFollowConnection = {
  __typename: "ModelUserFollowConnection",
  items:  Array<UserFollow | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type UserFollow = {
  __typename: "UserFollow",
  id: string,
  followerID: string,
  followeeID: string,
  Follower?: User | null,
  Followee?: User | null,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
  owner?: string | null,
};
// ...
export type CreateUserFollowInput = {
  id?: string | null,
  followerID: string,
  followeeID: string,
  _version?: number | null,
};

export type ModelUserFollowConditionInput = {
  followerID?: ModelIDInput | null,
  followeeID?: ModelIDInput | null,
  and?: Array< ModelUserFollowConditionInput | null > | null,
  or?: Array< ModelUserFollowConditionInput | null > | null,
  not?: ModelUserFollowConditionInput | null,
};

export type UpdateUserFollowInput = {
  id: string,
  followerID?: string | null,
  followeeID?: string | null,
  _version?: number | null,
};

export type DeleteUserFollowInput = {
  id: string,
  _version?: number | null,
};
// ...
export type ModelIDKeyConditionInput = {
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
};

export enum ModelSortDirection {
  ASC = "ASC",
  DESC = "DESC",
}
// ...
export type ModelUserFollowFilterInput = {
  id?: ModelIDInput | null,
  followerID?: ModelIDInput | null,
  followeeID?: ModelIDInput | null,
  and?: Array< ModelUserFollowFilterInput | null > | null,
  or?: Array< ModelUserFollowFilterInput | null > | null,
  not?: ModelUserFollowFilterInput | null,
};
// ...
export type CreateUserFollowMutationVariables = {
  input: CreateUserFollowInput,
  condition?: ModelUserFollowConditionInput | null,
};

export type CreateUserFollowMutation = {
  createUserFollow?:  {
    __typename: "UserFollow",
    id: string,
    followerID: string,
    followeeID: string,
    Follower?:  {
      __typename: "User",
      id: string,
      name: string,
      email?: string | null,
      username?: string | null,
      bio?: string | null,
      website?: string | null,
      nofPosts: number,
      nofFollowers: number,
      nofFollowings: number,
      image?: string | null,
      Posts?:  {
        __typename: "ModelPostConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Comments?:  {
        __typename: "ModelCommentConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Likes?:  {
        __typename: "ModelLikeConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Followers?:  {
        __typename: "ModelUserFollowConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Followings?:  {
        __typename: "ModelUserFollowConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
      owner?: string | null,
    } | null,
    Followee?:  {
      __typename: "User",
      id: string,
      name: string,
      email?: string | null,
      username?: string | null,
      bio?: string | null,
      website?: string | null,
      nofPosts: number,
      nofFollowers: number,
      nofFollowings: number,
      image?: string | null,
      Posts?:  {
        __typename: "ModelPostConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Comments?:  {
        __typename: "ModelCommentConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Likes?:  {
        __typename: "ModelLikeConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Followers?:  {
        __typename: "ModelUserFollowConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Followings?:  {
        __typename: "ModelUserFollowConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
      owner?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
    owner?: string | null,
  } | null,
};
// ...
export type DeleteUserFollowMutationVariables = {
  input: DeleteUserFollowInput,
  condition?: ModelUserFollowConditionInput | null,
};

export type DeleteUserFollowMutation = {
  deleteUserFollow?:  {
    __typename: "UserFollow",
    id: string,
    followerID: string,
    followeeID: string,
    Follower?:  {
      __typename: "User",
      id: string,
      name: string,
      email?: string | null,
      username?: string | null,
      bio?: string | null,
      website?: string | null,
      nofPosts: number,
      nofFollowers: number,
      nofFollowings: number,
      image?: string | null,
      Posts?:  {
        __typename: "ModelPostConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Comments?:  {
        __typename: "ModelCommentConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Likes?:  {
        __typename: "ModelLikeConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Followers?:  {
        __typename: "ModelUserFollowConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Followings?:  {
        __typename: "ModelUserFollowConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
      owner?: string | null,
    } | null,
    Followee?:  {
      __typename: "User",
      id: string,
      name: string,
      email?: string | null,
      username?: string | null,
      bio?: string | null,
      website?: string | null,
      nofPosts: number,
      nofFollowers: number,
      nofFollowings: number,
      image?: string | null,
      Posts?:  {
        __typename: "ModelPostConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Comments?:  {
        __typename: "ModelCommentConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Likes?:  {
        __typename: "ModelLikeConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Followers?:  {
        __typename: "ModelUserFollowConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Followings?:  {
        __typename: "ModelUserFollowConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
      owner?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
    owner?: string | null,
  } | null,
};
// ...
export type UserFollowingsQueryVariables = {
  followerID: string,
  followeeID?: ModelIDKeyConditionInput | null,
  sortDirection?: ModelSortDirection | null,
  filter?: ModelUserFollowFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type UserFollowingsQuery = {
  userFollowings?:  {
    __typename: "ModelUserFollowConnection",
    items:  Array< {
      __typename: "UserFollow",
      id: string,
      followerID: string,
      followeeID: string,
      Follower?:  {
        __typename: "User",
        id: string,
        name: string,
        email?: string | null,
        username?: string | null,
        bio?: string | null,
        website?: string | null,
        nofPosts: number,
        nofFollowers: number,
        nofFollowings: number,
        image?: string | null,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
        owner?: string | null,
      } | null,
      Followee?:  {
        __typename: "User",
        id: string,
        name: string,
        email?: string | null,
        username?: string | null,
        bio?: string | null,
        website?: string | null,
        nofPosts: number,
        nofFollowers: number,
        nofFollowings: number,
        image?: string | null,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
        owner?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
      owner?: string | null,
    } | null >,
    nextToken?: string | null,
    startedAt?: number | null,
  } | null,
};